# perf: memoise DataSerializer.read_file results