--- a/src/syrupy/extensions/amber/serializer.py
+++ b/src/syrupy/extensions/amber/serializer.py
@@ -1,3 +1,4 @@
+import functools
 import os
 from types import GeneratorType
 from typing import (
@@ -12,7 +13,6 @@ from typing import (
     Tuple,
     Union,
 )
-import functools
 
 from syrupy.constants import SYMBOL_ELLIPSIS
 from syrupy.data import (
@@ -80,7 +80,7 @@ class DataSerializer:
                     f.write(f"\n{cls._marker_divider}\n")
 
     @classmethod
-    @functools.lru_cache
+    @functools.lru_cache()
     def read_file(cls, filepath: str) -> "SnapshotFossil":
         """
         Read the raw snapshot data (str) from the snapshot file into a dict